--- Save.cpp
@@ -1,66 +1,0 @@
-#include "Save.h"
-#include <string>
-#include <windows.h>
-#include <iostream>
-
-bool File::SetLWTime()
-{
-	HANDLE hFile = CreateFile(path.c_str(), GENERIC_READ, FILE_SHARE_READ, NULL,
-		OPEN_ALWAYS, 0, NULL);
-
-	if (hFile == INVALID_HANDLE_VALUE)
-	{
-		std::cout << "CreateFile failed with " << GetLastError() << std::endl;
-		return 0;
-	}
-
-	FILETIME ftCreate, ftAccess, ftWrite;
-	SYSTEMTIME stUTC, stLocal;
-	DWORD dwRet;
-	double dblt;
-
-	// Retrieve the file times for the file.
-	if (!GetFileTime(hFile, &ftCreate, &ftAccess, &ftWrite))
-		return 0;
-
-	// Convert the last-write time to local time.
-	FileTimeToSystemTime(&ftWrite, &stUTC);
-	SystemTimeToVariantTime(&stUTC, &dblt);
-
-	CloseHandle(hFile);
-	time = dblt;
-
-	return 1;
-}
-
-double File::GetLWTime() const
-{
-	return time;
-}
-
-bool File::SetPathAndName(std::wstring dir)
-{
-	path = dir;
-	std::size_t i = dir.find_last_of(L"/\\");
-	if (i >= 0)
-	{
-		name = dir.substr(i + 1); // Name
-		return 1;
-	}
-	else
-	{
-		std::cout << "Your path is probably broken" << std::endl;
-		return 0;
-	}
-}
-
-std::wstring File::GetPath() const
-{
-	return path;
-}
-
-std::wstring File::GetName() const
-{
-	return name;
-}
-
